added tab-nav, dashboard component, refactor
