+++ packages/webapp/src/components/Locations.scss
@@ -3,119 +3,7 @@
 .locationPageContainer {
 	display: flex;
 	flex-direction: column;
-	min-height: 100vh;
-}
-
-.locationHeaderWrapper {
-	display: flex;
-	background-color: $ms-color-themePrimary;
-	height: 30px;
-	color: white;
-	align-items: center;
-	padding: 10px 20px;
-	/*min-width: 1360px;*/
-}
-
-.headerContainer {
-	font-weight: 600;
-	align-items: center;
-	display: flex;
-	margin-left: -10px;
-}
-
-.headerTitle {
-	margin-left: 10px;
-}
-
-.headerPersona {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	margin-left: auto;
-}
-
-.headerUsername {
-	font-size: 14px;
-	margin-right: 10px;
-	margin-left: 10px;
-}
-
-.footerPrivacy {
-	padding: 10px;
-	display: flex;
-
-	div {
-		font-size: 12px;
-	}
-}
-
-.locationBodyWrapper {
-	display: flex;
-	flex-direction: row;
-	flex: 1;
-	/*    min-width: 1400px;
-*/
-}
-
-.locationBodyLeft {
-	display: flex;
-	flex-direction: column;
-	margin: 0 40px;
-	flex: 1;
-	min-width: 900px;
-}
-
-.locationBodyRight {
-	display: flex;
-	flex-direction: column;
-	background-color: $ms-color-gray20;
-	width: 500px;
-}
-
-.bodyContainer {
-	display: flex;
-	flex-direction: column;
-	flex: 1;
-}
-
-.bodyHeader {
-	padding: 50px 0;
-	display: flex;
-	flex-direction: row;
-	justify-content: space-between;
-
-	&.subContent {
-		justify-content: unset;
-	}
-}
-
-.bodyHeaderTitle {
-	display: flex;
-	flex-direction: column;
-
-	.mainTitle {
-		font-size: 32px;
-	}
-
-	.breadCrumbs {
-		font-size: 14px;
-		margin-bottom: 10px;
-
-		.crumbLink {
-			font-size: 14px;
-			cursor: pointer;
-			color: $ms-color-themePrimary;
-		}
-	}
-}
-
-.bodyContent {
-	section {
-		/*box-shadow: $ms-depth-shadow-4;*/
-		padding: 10px;
-		margin-bottom: 20px;
-		border-radius: 2px;
-	}
+	min-height: calc(100vh - 130px);
 }
 
 .newLocationPanel {
